# Conflicts:
#	panel/lab/components/toolbar/3_writer/index.vue
#	panel/package-lock.json
#	panel/package.json
#	panel/src/components/Dropdowns/DropdownContent.vue
#	panel/src/components/Forms/Blocks/Block.vue
#	panel/src/components/Forms/Blocks/BlockType.vue
#	panel/src/components/Forms/Blocks/Blocks.vue
#	panel/src/components/Forms/Blocks/Elements/BlockFigure.vue
#	panel/src/components/Forms/Blocks/Types/Gallery.vue
#	panel/src/components/Forms/Blocks/Types/Heading.vue
#	panel/src/components/Forms/Blocks/Types/List.vue
#	panel/src/components/Forms/Blocks/Types/Text.vue
#	panel/src/components/Forms/Field/CheckboxesField.vue
#	panel/src/components/Forms/Field/ColorField.vue
#	panel/src/components/Forms/Field/DateField.vue
#	panel/src/components/Forms/Field/EmailField.vue
#	panel/src/components/Forms/Field/LinkField.vue
#	panel/src/components/Forms/Field/ListField.vue
#	panel/src/components/Forms/Field/MultiselectField.vue
#	panel/src/components/Forms/Field/NumberField.vue
#	panel/src/components/Forms/Field/PasswordField.vue
#	panel/src/components/Forms/Field/RadioField.vue
#	panel/src/components/Forms/Field/RangeField.vue
#	panel/src/components/Forms/Field/SelectField.vue
#	panel/src/components/Forms/Field/SlugField.vue
#	panel/src/components/Forms/Field/StructureField.vue
#	panel/src/components/Forms/Field/TagsField.vue
#	panel/src/components/Forms/Field/TelField.vue
#	panel/src/components/Forms/Field/TextField.vue
#	panel/src/components/Forms/Field/TextareaField.vue
#	panel/src/components/Forms/Field/TimeField.vue
#	panel/src/components/Forms/Field/ToggleField.vue
#	panel/src/components/Forms/Field/TogglesField.vue
#	panel/src/components/Forms/Field/UrlField.vue
#	panel/src/components/Forms/Field/WriterField.vue
#	panel/src/components/Forms/Input.vue
#	panel/src/components/Forms/Input/ColoroptionsInput.vue
#	panel/src/components/Forms/Input/DateInput.vue
#	panel/src/components/Forms/Input/EmailInput.vue
#	panel/src/components/Forms/Input/RadioInput.vue
#	panel/src/components/Forms/Input/RangeInput.vue
#	panel/src/components/Forms/Input/SelectInput.vue
#	panel/src/components/Forms/Input/SlugInput.vue
#	panel/src/components/Forms/Input/TextInput.vue
#	panel/src/components/Forms/Input/UrlInput.vue
#	panel/src/components/Forms/Input/WriterInput.vue
#	panel/src/components/Forms/Layouts/Layouts.vue
#	panel/src/components/Forms/Writer/Toolbar.vue
#	panel/src/components/Forms/Writer/Writer.vue
#	panel/src/components/Forms/Writer/index.js
#	panel/src/components/Lab/FieldExamples.vue
#	panel/src/components/Lab/FieldPreviewExample.vue
#	panel/src/components/Lab/InputExamples.vue
#	panel/src/components/Lab/InputboxExamples.vue
#	panel/src/components/Lab/OptionsFieldExamples.vue
#	panel/src/components/Lab/OptionsInputExamples.vue
#	panel/src/components/Lab/OptionsInputboxExamples.vue
#	panel/src/components/Navigation/Pagination.vue
#	panel/src/components/Text/Label.vue

+++ config/api/routes/users.php
@@ -1,11 +1,14 @@
 <?php
 
+use Kirby\Exception\Exception;
 use Kirby\Filesystem\F;
+use Kirby\Toolkit\Str;
 
 /**
  * User Routes
  */
 return [
+	// @codeCoverageIgnoreStart
 	[
 		'pattern' => 'users',
 		'method'  => 'GET',
@@ -79,10 +82,27 @@ return [
 		],
 		'method'  => 'POST',
 		'action'  => function (string $id) {
-			$this->user($id)->avatar()?->delete();
-
 			return $this->upload(
 				function ($source, $filename) use ($id) {
+					$type = F::type($filename);
+					if ($type !== 'image') {
+						throw new Exception([
+							'key'  => 'file.type.invalid',
+							'data' => compact('type')
+						]);
+					}
+
+					$mime = F::mime($source);
+					if (Str::startsWith($mime, 'image/') !== true) {
+						throw new Exception([
+							'key'  => 'file.mime.invalid',
+							'data' => compact('mime')
+						]);
+					}
+
+					// delete the old avatar
+					$this->user($id)->avatar()?->delete();
+
 					$props = [
 						'filename' => 'profile.' . F::extension($filename),
 						'template' => 'avatar',
@@ -188,6 +208,16 @@ return [
 	],
 	[
 		'pattern' => [
+			'(account)/fields/(:any)/(:all?)',
+			'users/(:any)/fields/(:any)/(:all?)',
+		],
+		'method'  => 'ALL',
+		'action'  => function (string $id, string $fieldName, string|null $path = null) {
+			return $this->fieldApi($this->user($id), $fieldName, $path);
+		}
+	],
+	[
+		'pattern' => [
 			'(account)/sections/(:any)',
 			'users/(:any)/sections/(:any)',
 		],
@@ -200,12 +230,13 @@ return [
 	],
 	[
 		'pattern' => [
-			'(account)/fields/(:any)/(:all?)',
-			'users/(:any)/fields/(:any)/(:all?)',
+			'(account)/sections/(:any)/(:all?)',
+			'users/(:any)/sections/(:any)/(:all?)',
 		],
 		'method'  => 'ALL',
-		'action'  => function (string $id, string $fieldName, string $path = null) {
-			return $this->fieldApi($this->user($id), $fieldName, $path);
+		'action'  => function (string $id, string $sectionName, string|null $path = null) {
+			return $this->sectionApi($this->user($id), $sectionName, $path);
 		}
 	],
+	// @codeCoverageIgnoreEnd
 ];
